# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "W03 CRUD API",
    "description": "API with two collections (items and categories). Full CRUD, validation, error handling, and GitHub OAuth authentication.",
    "version": "2.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://w03-crud-api-vmrb.onrender.com",
      "description": "Production server on Render (Original)"
    },
    {
      "url": "https://w03-crud-api-1.onrender.com",
      "description": "Production server on Render (Instance 2 - Main)"
    },
    {
      "url": "http://localhost:8083",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "GitHub OAuth2 Authentication - Click 'Authorize' button below and login with GitHub",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://github.com/login/oauth/authorize",
            "tokenUrl": "https://github.com/login/oauth/access_token",
            "scopes": {
              "read:user": "Read user information",
              "user:email": "Read user email addresses"
            }
          }
        }
      },
      "apiKey": {
        "type": "apiKey",
        "name": "API_SECRET",
        "in": "header",
        "description": "API Secret key for additional security"
      }
    },
...
  },
  "security": [
    {
      "oauth2": ["read:user", "user:email"]
    }
  ],
  "paths": {
...
    "/auth/github": {
      "get": {
        "summary": "Authenticate with GitHub",
        "tags": ["Authentication"],
        "description": "Redirects to GitHub for OAuth authentication. Uses Client ID: Ov23liVpFGbPyMJFrYYM",
        "externalDocs": {
          "description": "GitHub OAuth Documentation",
          "url": "https://docs.github.com/en/apps/oauth-apps/building-oauth-apps/authorizing-oauth-apps"
        },
        "responses": {
          "302": {
            "description": "Redirect to GitHub login page",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "GitHub OAuth authorization URL"
              }
            }
          },
          "500": {
            "description": "Authentication error"
          }
        }
      }
    },
...
  },
...
}